Added flags to d00/ex00

+++ d00/ex00/megaphone.cpp
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   megaphone.cpp                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ikozlov <ikozlov@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ivankozlov <ivankozlov@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/24 19:10:57 by ikozlov           #+#    #+#             */
-/*   Updated: 2018/06/24 19:32:28 by ikozlov          ###   ########.fr       */
+/*   Updated: 2019/04/29 17:54:51 by ivankozlov       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <iostream>
 
 char *toUpper(char *s) {
-	for (int i = 0; i < strlen(s); i++)
+	for (size_t i = 0; i < strlen(s); i++)
 		s[i] = toupper(s[i]);
 	return (s);
 }
